Restructured src/lib for clarity and updated imports
