--- src/lib/app/routes.ts
@@ -1,18 +1,0 @@
-export const routes = {
-  home: {
-    href: '/',
-    label: 'Home',
-  },
-  pokedex: {
-    href: '/pokedex',
-    label: 'Pokédex',
-  },
-  types: {
-    href: '/types',
-    label: 'Types',
-  },
-  favourites: {
-    href: '/favourites',
-    label: 'Favourites',
-  },
-} as const;
